+++ R/require.R
@@ -87,11 +87,93 @@ require_module <- function(module, options = NULL, version = NULL,
   reticulate::py_require(requirement, python_version = python_version)
 }
 
+#' Require lamindb
+#'
+#' Require the lamindb Python module
+#'
+#' @param silent Whether to suppress messages showing what has been required
+#'
+#' @noRd
+#'
+#' @details
+#' Functions requiring the `lamindb` Python module should make sure this is
+#' called before attempting to use it (either directly, or via
+#' `import_module()`).
 require_lamindb <- function(silent = FALSE) {
-  require_module(
-    "lamindb",
-    version = ">=1.2",
-    python_version = ">=3.10,<3.14",
-    silent = silent
-  )
+  if (reticulate::py_available() && reticulate::py_module_available("lamindb")) {
+    return(invisible(NULL))
+  }
+
+  laminr_lamindb_version <- trimws(tolower(Sys.getenv("LAMINR_LAMINDB_VERSION")))
+  laminr_lamindb_options <- Sys.getenv("LAMINR_LAMINDB_OPTIONS")
+  if (laminr_lamindb_options != "") {
+    laminr_lamindb_options <- trimws(unlist(strsplit(laminr_lamindb_options, ",")))
+    if (!isTRUE(silent)) {
+      cli::cli_alert_info(
+        "Requiring {.pkg lamindb} options {.val {laminr_lamindb_options}}"
+      )
+    }
+  } else {
+    laminr_lamindb_options <- NULL
+  }
+
+  if (laminr_lamindb_version %in% c("release", "latest", "")) {
+    require_module(
+      "lamindb",
+      options = laminr_lamindb_options,
+      version = ">=1.2",
+      python_version = ">=3.10,<3.14",
+      silent = silent
+    )
+  } else if (laminr_lamindb_version %in% c("github", "devel")) {
+    if (!isTRUE(silent)) {
+      cli::cli_alert_info(
+        "Requiring the development version of {.pkg lamindb}"
+      )
+    }
+
+    # Also require devel versions of other lamin packages
+    require_module(
+      "lamindb_setup",
+      options = "aws",
+      source = "git+https://github.com/laminlabs/lamindb-setup.git",
+      silent = silent
+    )
+    require_module(
+      "lamin_utils",
+      source = "git+https://github.com/laminlabs/lamin-utils.git",
+      silent = silent
+    )
+    require_module(
+      "lamin_cli",
+      source = "git+https://github.com/laminlabs/lamin-cli.git",
+      silent = silent
+    )
+    require_module(
+      "lamindb",
+      options = laminr_lamindb_options,
+      source = "git+https://github.com/laminlabs/lamindb.git",
+      silent = silent
+    )
+  } else {
+    # Remove leading v from version string
+    laminr_lamindb_version <- gsub("^v", "", laminr_lamindb_version)
+    # Assume an exact version if string starts with a number
+    if (grepl("^[0-9]", laminr_lamindb_version)) {
+      laminr_lamindb_version <- paste0("==", laminr_lamindb_version)
+    }
+
+    if (!isTRUE(silent)) {
+      cli::cli_alert_info(
+        "Requiring {.pkg lamindb} version {.val {laminr_lamindb_version}}"
+      )
+    }
+
+    require_module(
+      "lamindb",
+      options = laminr_lamindb_options,
+      version = laminr_lamindb_version,
+      silent = silent
+    )
+  }
 }